refactor(sidecar): chain_io module
