+++ bolt-sidecar/src/chain_io/manager.rs
@@ -2,7 +2,7 @@ use std::str::FromStr;
 
 use alloy::{
     contract::Error as ContractError,
-    primitives::{Address, Bytes, B256, B512},
+    primitives::{Address, Bytes},
     providers::{ProviderBuilder, RootProvider},
     sol,
     sol_types::{Error as SolError, SolInterface},
@@ -11,33 +11,23 @@ use alloy::{
 use ethereum_consensus::primitives::BlsPublicKey;
 use eyre::bail;
 use reqwest::{Client, Url};
-use reth_primitives::keccak256;
 use serde::Serialize;
 
 use BoltManagerContract::{BoltManagerContractErrors, BoltManagerContractInstance, ProposerStatus};
 
 use crate::config::chain::Chain;
 
+use super::utils;
+
 /// A wrapper over a BoltManagerContract that exposes various utility methods.
 #[derive(Debug, Clone)]
 pub struct BoltManager(BoltManagerContractInstance<Http<Client>, RootProvider<Http<Client>>>);
 
 impl BoltManager {
-    /// Returns the address of the canonical BoltManager contract for a given chain, if present
-    pub fn address(chain: Chain) -> Option<Address> {
-        match chain {
-            Chain::Holesky => Some(
-                Address::from_str("0x440202829b493F9FF43E730EB5e8379EEa3678CF")
-                    .expect("valid address"),
-            ),
-            _ => None,
-        }
-    }
-
     /// Creates a new BoltRegistry instance. Returns `None` if a canonical BoltManager contract is
     /// not deployed on such chain.
     pub fn from_chain<U: Into<Url>>(execution_client_url: U, chain: Chain) -> Option<Self> {
-        let address = Self::address(chain)?;
+        let address = chain.manager_address()?;
         Some(Self::from_address(execution_client_url, address))
     }
 
@@ -66,7 +56,7 @@ impl BoltManager {
         &self,
         keys: &[BlsPublicKey],
     ) -> eyre::Result<Vec<ProposerStatus>> {
-        let hashes = BoltValidators::pubkey_hashes(keys);
+        let hashes = utils::pubkey_hashes(keys);
 
         let returndata = self.0.getProposerStatuses(hashes).call().await;
 
@@ -110,37 +100,8 @@ impl BoltManager {
     }
 }
 
-/// Utility functions related to the BoltValidators contract
-pub struct BoltValidators;
-
-impl BoltValidators {
-    /// Hash the public keys of the proposers. This follows the same
-    /// implementation done on-chain in the BoltValidators contract.
-    pub fn pubkey_hashes(keys: &[BlsPublicKey]) -> Vec<B256> {
-        keys.iter().map(Self::pubkey_hash).collect()
-    }
-
-    /// Hash the public key of the proposer. This follows the same
-    /// implementation done on-chain in the BoltValidators contract.
-    pub fn pubkey_hash(key: &BlsPublicKey) -> B256 {
-        let digest = Self::pubkey_hash_digest(key);
-        keccak256(digest)
-    }
-
-    fn pubkey_hash_digest(key: &BlsPublicKey) -> B512 {
-        let mut onchain_pubkey_repr = B512::ZERO;
-
-        // copy the pubkey bytes into the rightmost 48 bytes of the 512-bit buffer.
-        // the result should look like this:
-        //
-        // 0x00000000000000000000000000000000b427fd179b35ef085409e4a98fb3ab84ee29c689df5c64020eab0b20a4f91170f610177db172dc091682df627c9f4021
-        // |<---------- 16 bytes ---------->||<----------------------------------------- 48 bytes ----------------------------------------->|
-        onchain_pubkey_repr[16..].copy_from_slice(key);
-        onchain_pubkey_repr
-    }
-}
-
 sol! {
+    #[allow(missing_docs)]
     #[sol(rpc)]
     interface BoltManagerContract {
         #[derive(Debug, Default, Serialize)]
