+++ bolt-sidecar/src/chain_io/mod.rs
@@ -1,0 +1,5 @@
+/// Wrapper over the BoltManager contract
+pub mod manager;
+
+/// Utilities and functions used in the Bolt contracts
+pub mod utils;
