+++ bolt-sidecar/src/chain_io/utils.rs
@@ -1,0 +1,28 @@
+use alloy::primitives::{B256, B512};
+use ethereum_consensus::primitives::BlsPublicKey;
+use reth_primitives::keccak256;
+
+/// Hash the public keys of the proposers. This follows the same
+/// implementation done on-chain in the BoltValidators contract.
+pub fn pubkey_hashes(keys: &[BlsPublicKey]) -> Vec<B256> {
+    keys.iter().map(pubkey_hash).collect()
+}
+
+/// Hash the public key of the proposer. This follows the same
+/// implementation done on-chain in the BoltValidators contract.
+pub fn pubkey_hash(key: &BlsPublicKey) -> B256 {
+    let digest = pubkey_hash_digest(key);
+    keccak256(digest)
+}
+
+fn pubkey_hash_digest(key: &BlsPublicKey) -> B512 {
+    let mut onchain_pubkey_repr = B512::ZERO;
+
+    // copy the pubkey bytes into the rightmost 48 bytes of the 512-bit buffer.
+    // the result should look like this:
+    //
+    // 0x00000000000000000000000000000000b427fd179b35ef085409e4a98fb3ab84ee29c689df5c64020eab0b20a4f91170f610177db172dc091682df627c9f4021
+    // |<---------- 16 bytes ---------->||<----------------------------------------- 48 bytes ----------------------------------------->|
+    onchain_pubkey_repr[16..].copy_from_slice(key);
+    onchain_pubkey_repr
+}
