+++ bolt-sidecar/src/lib.rs
@@ -48,8 +48,8 @@ pub mod state;
 mod signer;
 pub use signer::{commit_boost::CommitBoostSigner, SignerBLS};
 
-mod onchain_registry;
-pub use onchain_registry::BoltManager;
+/// Utilities and contracts wrappers for interacting with the Bolt registry
+pub mod chain_io;
 
 /// Utilities for testing
 #[cfg(test)]
